% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddDist2Colony.R
\name{AddDist2Colony}
\alias{AddDist2Colony}
\title{AddDist2Colony Calculate the Distance from Points to the Many Colonies}
\usage{
AddDist2Colony(tracks, dataLat = "lat", dataLon = "lon", CaptureSitesData,
  SiteName = "SiteShort")
}
\arguments{
\item{tracks}{a data frame with with latitude and longitude and colony site names}

\item{dataLat}{quoted name of column in data that has latitude values}

\item{dataLon}{quoted name of column in data that has longitude values}

\item{CaptureSitesData}{data.frame with colony site names and colony lat long}

\item{SiteName}{a quoted string indicating what the column houses your SiteNames.
Must be the same across both your data and the CaptureSitesData}
}
\value{
vector of distances from each colony
}
\description{
AddDist2Colony Calculate the Distance from Points to the Many Colonies
}
\author{
Abram B. Fleishman <abram.fleishman AT sjsu.edu>
}
