% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cost_choice_index.R
\name{cost_choice_index}
\alias{cost_choice_index}
\title{Calculate cost choice index for a single point}
\usage{
cost_choice_index(wind.speed, bird.bearing.to, wind.bearing.to)
}
\arguments{
\item{wind.speed}{a single wind speed in meters per second}

\item{bird.bearing.to}{a single bering from the previous point to the current point along a tajectory(in degrees)}

\item{wind.bearing.to}{a single bering for the wind trajectory (in degrees)}
}
\description{
Calculate the cost choice index for a single point
}
\author{
Abram B. Fleishman \email{abram@conservationmetrics.com}

Caitlin Kroeger \email{ckroeger@ucsc.edu}
}
