% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DepthETOPO.R
\name{DepthETOPO}
\alias{DepthETOPO}
\title{DepthETOPO Compute the depth from Etopo1 is at 1 arc minute resolution.}
\usage{
DepthETOPO(tracks, dataLat = "Latitude", dataLon = "Longitude",
  left = -179, right = -164, top = 60, bottom = 50,
  resolution = 30, keep = F)
}
\arguments{
\item{tracks}{data.frame of data to be queried for  latand long values}

\item{dataLat}{quoted name of column in data that has latitude values}

\item{dataLon}{quoted name of column in data that has longitude values}

\item{left}{The left side of the bounding box of data}

\item{right}{The right side of the bounding box of data}

\item{top}{The top side of the bounding box of data}

\item{bottom}{The bottom side of the bounding box of data}

\item{resolution}{arc minutes res}

\item{keep}{save a file}
}
\value{
A vector of elevation/depth in meters from ETOPO1 world bathymatry at a resolution of 1 arc minute
}
\description{
It would be nice to find something higher resolution
}
\author{
Abram B. Fleishman \email{abram@conservationmetrics.com}
}
