% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dist2Colony.R
\name{Dist2Colony}
\alias{Dist2Colony}
\title{Dist2Colony Calculate the distance of a vector of points (lat,long) to a fixed point}
\usage{
Dist2Colony(tracks, dataLat = "lat", dataLon = "lon", ColonyLat,
  ColonyLong)
}
\arguments{
\item{tracks}{a dataframe}

\item{dataLat}{quoted latitude}

\item{dataLon}{quoted longitude}

\item{ColonyLat}{a latitude value of length = 1}

\item{ColonyLong}{a longitude value of length = 1}
}
\value{
a vector of distances from adjacent points in kilometers
}
\description{
Calculate the distance of a vector of points (lat,long) to a fixed point
}
\author{
Abram B. Fleishman <abram.fleishman AT sjsu.edu>
}
