% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpointDist.R
\name{distance}
\alias{distance}
\title{Distance (from argosfilter)}
\usage{
distance(lat1, lat2, lon1, lon2)
}
\arguments{
\item{lat1}{latitude of first point}

\item{lat2}{longtitude of first point}

\item{lon2}{lonitude of second point}
}
\description{
Distance (from argosfilter)
}
