% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpointDist.R
\name{distanceTrack}
\alias{distanceTrack}
\title{distanceTrack (from argosfilter)}
\usage{
distanceTrack(lat, lon)
}
\arguments{
\item{lat}{vector of latitudes}

\item{lon}{vector of longitudes}
}
\description{
distanceTrack (from argosfilter)
}
