% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gilmour_fedelity_index.R
\name{gilmour_fidelity}
\alias{gilmour_fidelity}
\title{Gilmour's Fidelity Index}
\usage{
gilmour_fidelity(
  dat,
  Longitude,
  Latitude,
  animal_id,
  trip_id,
  distance2colony,
  bear2col
)
}
\arguments{
\item{dat}{a data frame of tracking data with columns that include positions
(longitude/Latitude, Easting/Northing,etc), an animal ID, unique trip IDs
(unique to each animal/trip combo), a column with a distance to the colony,
and a column with a bearing from each point to the colony.}

\item{Longitude}{quoted column name for location in the x direction.  can be
longitude or easting or any other x coordinator}

\item{Latitude}{quoted column name for location in the y direction.  can be
latitude or northing or any other y coordinator}

\item{animal_id}{quoted column name of a unique ID field for each animal}

\item{trip_id}{quoted column name of a unique ID field for each animal/trip
(can be a combo of animalID and a trip number for that animal)}

\item{distance2colony}{quoted column name of a distance to the colony (units
do not matter)}

\item{bear2col}{quoted column name of a bearing in degrees from each point to
the colony}
}
\value{
a list of data.frames with the fidelity index for each trip
  combination (fidelity_index_trips ) and an average fidelity index for each animal (fidelity_index_animal)
}
\description{
Used for calculation of Fidelity Index as described by: Gilmour, ME, JA
Castillo-Guerrero, AB Fleishman, S Hernandez-Vazquez, HS Young, SA Shaffer.
2018. Plasticity of foraging behaviors in response to diverse environmental
conditions. Ecosphere 9(7):e02301. https://doi.org/10.1002/ecs2.2301 (open access)
}
\details{
This function compares maximum displacement (distance and angle) of two
separate trips from a point of origin. Multiple displacement values can be
compared. These displacement values are centered and scaled to create an
index of site fidelity that ranges -1 to 1.

The equations used in this script are modified from:
 Willis-Norton et al. 2015. Deep-Sea Research Part II: Topical Studies in Oceanography 113:260-267
 Hazen et al. 2016. Journal of Applied Ecology 54:1415-1428, and
 Shaffer et al. 2017. Movement Ecology 5:27 https://doi.org/10.1186/s40462-017-0118-9.
}
\author{
Morgan E. Gilmour \email{morgan.gilmour@bucknell.edu}

Abram B. Fleishman \email{abram@conservationmetrics.com}
}
