% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpointDist.R
\name{InterpointDist}
\alias{InterpointDist}
\title{InterpointDist Calculate the Distance between points on a track for each bird
Vector of distances between points}
\usage{
InterpointDist(tracks, ID = "CaptureNum", lat = "Latitude",
  lon = "Longitude")
}
\arguments{
\item{tracks}{data.frame of data to be queried for  lat and long values}

\item{ID}{quoted name of column in data that is and ID field}

\item{lat}{quoted name of column in data that has latitude values}

\item{lon}{quoted name of column in data that has longitude values}
}
\value{
A vector of distances in meters between adjacent points in an animal track
}
\description{
InterpointDist Calculate the Distance between points on a track for each bird
Vector of distances between points
}
\author{
Abram B. Fleishman <abram.fleishman AT sjsu.edu>
}
