% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpointTime.R
\name{InterpointTime}
\alias{InterpointTime}
\title{Calculate the time between points on a track for each bird
Vector of times between points}
\usage{
InterpointTime(tracks = tracks, ID = "File", DateTime = "DateTime")
}
\arguments{
\item{tracks}{data.frame of data to be queried for  latand long values}

\item{ID}{quoted name of column in data that is and ID field}

\item{DateTime}{quoted name of column in data that has DateTime values in the format YYYY-mm-dd HH:MM:SS}
}
\value{
A vector of times differnces in seconds between adjacent points in an animal track
}
\description{
Calculate the time between points on a track for each bird
Vector of times between points
}
\author{
Abram B. Fleishman <abram.fleishman AT sjsu.edu>
}
