% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/label_events.R
\name{label_events}
\alias{label_events}
\title{Number events (such as foraging events) by bird and trip}
\usage{
label_events(dat, BirdID = "CaptureID", TripID = "TripID",
  eventCol = "forage")
}
\arguments{
\item{dat}{data.frame of with a minimum of three columns with birdID, tripID
from the MakeTrip function and a event column}

\item{BirdID}{a unique ID for each bird}

\item{TripID}{a unique ID for each trip nested within each bird. I.e. birdA
has trips 1,2,3; BirdB has trips 1,2,3,4}

\item{eventCol}{a column with values indicating the presence of an event.
Must bein the format 0= no event, 1= event}
}
\value{
A new data frame with all original data plus two new columns:
  event_bird with numbered events 1:n_vents by bird, and event_trip with
  numbered events 1:n_events per trip
  @importFrom dplyr bind_rows
}
\description{
Sometimes it is useful to number events consecutively such as forageing
events for a seabird.  often times these events have multiple rows per
event/bout.  This function numbers these evens consecutively by bird and by
trip nested within bird.
}
\author{
Abram B. Fleishman <abram.fleishman AT sjsu.edu>
}
