% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MakeTrip.R
\name{MakeTrip}
\alias{MakeTrip}
\title{MakeTrip Function to make a TripNum variable}
\usage{
MakeTrip(
  tracks,
  ID = "File",
  DistCutOff = 10,
  Dist2Colony = "Dist2Colony",
  NumLocCut = 3
)
}
\arguments{
\item{tracks}{data.frame of data that you want to parse into trips}

\item{ID}{quoted name of column in data that is a unique key to individual
bird_tag_deployment combos. This is the File from the GypsyLocRead output}

\item{DistCutOff}{Distance in km to use as a cut off radius around the colony
to use to split the trips}

\item{Dist2Colony}{quoted name of column in data that has the distance in km
from each point to the colony}

\item{NumLocCut}{Minimum number of points to consider in a trip}
}
\value{
A new data frame with all original data plus two new columns: TripNum
  (consecutive trip number 0 = at colony) and ColonyMovement (in/out colony
  movements)
}
\description{
The objective of this function to to partition tracking data into "trips"
based on begining and returning to a fixed location.  It was designed for use
with GPS tracking data from seabirds that during the breeding season are
central place foragers, leaving from colony sites on food finding trips and
returning to the colony afterward.  This function relies on a column with a
disntance in km from each point along the track to the colony and a distance
cutoff.  It numbers each trip starting with the first point beyond the
distance cutoff and ending with the last point beyond the cutoff with a
sequential number starting with 1.  All points inside the cutoff are labled
0.  In the future it could be nice to have these labeled with sequetial
numbers as well.
}
\author{
Abram B. Fleishman \email{abram@conservationmetrics.com}
}
