% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tracks.R
\docType{data}
\name{tracks}
\alias{tracks}
\title{tracks from a red-legged kittiwake}
\format{A data frame with 2964 rows and 5 variables:
\describe{
  \item{DateTime}{datetime in yyy-mm-dd hh:mm:ss}
  \item{Longitude}{longitude}
  \item{Latitude}{latitude}
  \item{CaptureID}{bird capture id}
  \item{Dist2Colony}{distance in km from the colony}
}}
\usage{
tracks
}
\description{
this dataset is a not big...
}
\author{
Abram B. Fleishman <abram.fleishman AT sjsu.edu>
}
\keyword{datasets}
