% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/true_cost.R
\name{true_cost}
\alias{true_cost}
\title{Calculate true cost of flight}
\usage{
true_cost(wind.speed, bird.bearing.to, wind.bearing.to)
}
\arguments{
\item{wind.speed}{a vector of wind speed in meters per second}

\item{bird.bearing.to}{a vector bering from the previous point to the current point along a tajectory(in degrees)}

\item{wind.bearing.to}{a vector bering for the wind trajectory (in degrees)}
}
\description{
Calculate the true cost of flight with the wind
}
\author{
Abram B. Fleishman \email{abram@conservationmetrics.com}

Caitlin Kroeger \email{ckroeger@ucsc.edu}
}
