% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TurningAngle.R
\name{TurningAngle}
\alias{TurningAngle}
\title{Calculate the turning angles between points on a track for each bird}
\usage{
TurningAngle(tracks, ID = "File", Bearing = "PointBearing")
}
\arguments{
\item{tracks}{data.frame of data to be queried for  latand long values}

\item{ID}{quoted name of column in data that is and ID field}

\item{Bearing}{quoted name of column in data that has the bearing between points in an animal track}
}
\value{
A vector of turning angles (degrees) between adjacent points in an animal track
}
\description{
Calculate the turning angles between points on a track for each bird
}
\author{
Abram B. Fleishman <abram.fleishman AT sjsu.edu>
}
