% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AddDist2Colony.R
\name{AddDist2Colony}
\alias{AddDist2Colony}
\title{AddDist2Colony Calculate the Distance from Points to Many Colonies}
\usage{
AddDist2Colony(
  tracks,
  dataLat = "lat",
  dataLon = "lon",
  CaptureSitesData,
  SiteName = "SiteShort",
  capLat = "lat",
  capLon = "lon"
)
}
\arguments{
\item{tracks}{a data frame with with latitude and longitude and colony site names}

\item{dataLat}{quoted name of column in data that has latitude values}

\item{dataLon}{quoted name of column in data that has longitude values}

\item{CaptureSitesData}{data.frame with colony site names and colony lat long}

\item{SiteName}{a quoted string indicating what the column houses your SiteNames.
Must be the same across both your data and the CaptureSitesData}

\item{capLat}{quoted name of column in CaptureSitesData that has latitude values}

\item{capLon}{quoted name of column in CaptureSitesData that has longitude values}
}
\value{
vector of distances from each colony
}
\description{
AddDist2Colony takes tracks with lat/lon positions and and
}
\author{
Abram B. Fleishman \email{abram@conservationmetrics.com}
}
