% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BearingFromPoint.R
\name{BearingFromPoint}
\alias{BearingFromPoint}
\title{BearingFromPoint Calculate bearing between two points along a vector}
\usage{
BearingFromPoint(tracks, ID = "File", lat = "Latitude",
  lon = "Longitude")
}
\arguments{
\item{tracks}{a dataframe}

\item{ID}{a quoted string indicating the column name for a unique ID key}

\item{lat}{a quoted string indicating the column name for longitude values}

\item{lon}{a quoted string indicating the column name for latitude values}
}
\value{
a vector of bearings from adjacent points
}
\description{
BearingFromPoint Calculate bearing between two points along a vector
}
\author{
Abram B. Fleishman \email{abram@conservationmetrics.com}
}
