% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InOutPoints.R
\name{InOutPoints}
\alias{InOutPoints}
\title{Inbound/Outbound flight leg identification}
\usage{
InOutPoints(tracks = tracks, CaptureID = "CaptureID",
  TripID = "TripID", dist2colony = "dist2colony", lag = 1,
  nPointsToSmooth = 10, minDist2Col = 5, Plot = F, Lon = "lon",
  Lat = "lat", pdfName = "inout_plots.pdf")
}
\arguments{
\item{tracks}{dataframe with tracking data sorted by ID, tripID, datetime}

\item{CaptureID}{column name holding animal ID}

\item{TripID}{column name holding Trip ID}

\item{dist2colony}{column name holding distances to the colony of each point}

\item{lag}{is the number of points to skip to caculate the change in distance. e.g. if lag=3, the diff between i and i-3 will be calculated}

\item{nPointsToSmooth}{number of points to smooth to determine legs}

\item{minDist2Col}{buffer in km around colony that will removed to avoid false out/in leg identification}

\item{Plot}{True/False whether to plot each trip (good for figuring out the right parameters)}

\item{Lon}{column name holding Longitude (x-coord) (only needed if Plot==T)}

\item{Lat}{column name holding Latitude (y coord) (only needed if Plot==T)}

\item{pdfName}{fullpath to pdf that will be saved (only needed if Plot==T)}
}
\description{
This functions intention is to partition a track into an inbound and outbound
leg demarkated when the bird first stops moving away from the colony. The
first stopping point is identified by taking the change in dist2col first
dirivative?, smoothing it with a left aligned rolling mean, and rounding to 1
digit.  The first point where the bird is not changing the distance to the
colony is the end of the outbound leg, while the last point where the birds is
not getting close to the colony is the inbound leg.
}
\details{
METHODS: Trip departure and return segments were identified by calculating ∆d
, the change in distance from the colony, lagged by lag gps points, and smoothed
by calculating a 10 point left aligned rolling mean. These values were rounded
to 1 significant digit. Trips were segmented by assigning points between the
beginning of a trip and the first point ∆d which equaled zero as the “out” leg
and the last point ∆d that equaled zero until the final point of the trip as
the “in” leg. The island was buffered so that the first and last points ∆d
could not be within 10 km of the island. If there was no point on the trip
that ∆d reached zero, i.e. the bird continuously moved away from the colony
until it began to return, the trip was split at the apex point where the bird
was furthest from the colony. This method effectively finds the first and last
points that the bird ceases to move away from the colony over a 15-minute
period.
}
