% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/InterpointTime.R
\name{InterpointTime}
\alias{InterpointTime}
\title{InterpointTime Calculate the time between points on a track for each bird}
\usage{
InterpointTime(tracks, ID = "File", DateTime = "DateTime")
}
\arguments{
\item{tracks}{data.frame of data to be queried for lat and long values}

\item{ID}{quoted name of column in data that is and ID field}

\item{DateTime}{quoted name of column in data that has DateTime values in the
format YYYY-mm-dd HH:MM:SS}
}
\value{
A vector of times differences in seconds between adjacent points in
  an animal track
}
\description{
Vector of times between points
}
\author{
Abram B. Fleishman \email{abram@conservationmetrics.com}
}
