% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Speed.R
\name{Speed}
\alias{Speed}
\title{Calculate the instantanious speed at each point along on a track for each bird
Vector of times between points}
\usage{
Speed(Dist = tracks$Dist2Point, Time = tracks$InterPointTime)
}
\arguments{
\item{Dist}{vector of distances between points}

\item{Time}{vector of times between points}
}
\value{
A vector of speeds (km/h) between adjacent points in an animal track
}
\description{
Calculate the instantanious speed at each point along on a track for each bird
Vector of times between points
}
\author{
Abram B. Fleishman <abram.fleishman AT sjsu.edu>
}
